Move utils.h and utils.c to util directory.

+++ media/util/utils.h
@@ -3,6 +3,6 @@
 
 #include <time.h>
 
-int64_t timespecDiff(struct timespec *timeA_p, struct timespec *timeB_p);
+int timespecDiff(struct timespec *timeA_p, struct timespec *timeB_p);
 
 #endif /* __UTILS_H__ */
